feat: 实现分片上传功能

+++ src/main/java/com/zoi/drive/entity/dto/UserFile.java
@@ -42,7 +42,7 @@ public class UserFile implements Serializable {
 
     private String type;
 
-    private Double size;
+    private long size;
 
     private String hash;
